+++ package.json
@@ -56,6 +56,7 @@
     "react-hot-loader": "3.0.0-beta.6",
     "react-portal": "^3.1.0",
     "react-redux": "^5.0.5",
+    "react-sizeme": "^2.3.6",
     "react-swipeable": "^4.0.1",
     "react-tap-event-plugin": "^3.0.2",
     "react-templates": "^0.6.1",
