+++ pages/Cart/components/PaymentBar/components/Content/components/CheckoutButton/spec.jsx
