+++ pages/Cart/components/PaymentBar/components/Content/components/SubTotalLabel/connector.js
